Factored out position structs
